add auth for user and role tests

+++ auth/tests/test_roles.py
@@ -12,8 +12,8 @@ def user_role_id():
     return role_id
 
 
-def test_get_all_roles(client):
-    response = client.get("/auth/roles")
+def test_get_all_roles(client, auth_as_admin):
+    response = client.get("/auth/roles", headers=auth_as_admin)
 
     assert response.status_code == HTTPStatus.OK
     for role in response.json:
@@ -25,13 +25,13 @@ def test_get_all_roles(client):
     assert set(role["name"] for role in response.json) == {"admin", "subscriber", "user"}
 
 
-def test_create_role(client):
-    response = client.post("/auth/roles", json={"name": "test"})
+def test_create_role(client, auth_as_admin):
+    response = client.post("/auth/roles", json={"name": "test"}, headers=auth_as_admin)
     assert response.status_code == HTTPStatus.CREATED
     assert response.json.get("name") == "test"
 
     role_id = response.json.get("id")
-    response = client.get(f"/auth/roles/{role_id}")
+    response = client.get(f"/auth/roles/{role_id}", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.OK
     assert response.json.get("name") == "test"
 
@@ -44,14 +44,14 @@ def test_create_role(client):
         ({"name": "test", "excess_key": "value"}, HTTPStatus.CREATED),
     ],
 )
-def test_create_role_errors(query, status_code, client):
-    response = client.post("/auth/roles", json=query)
+def test_create_role_errors(query, status_code, client, auth_as_admin):
+    response = client.post("/auth/roles", json=query, headers=auth_as_admin)
 
     assert response.status_code == status_code
 
 
-def test_get_role(client, user_role_id):
-    response = client.get(f"/auth/roles/{user_role_id}")
+def test_get_role(client, user_role_id, auth_as_admin):
+    response = client.get(f"/auth/roles/{user_role_id}", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.OK
     assert response.json.get("name") == "user"
 
@@ -63,16 +63,16 @@ def test_get_role(client, user_role_id):
         ("7f32cd4a-7981-436d-bba7-78169acbbb5d", HTTPStatus.NOT_FOUND),
     ],
 )
-def test_get_role_errors(role_id, status_code, client):
-    response = client.get(f"/auth/roles/{role_id}")
+def test_get_role_errors(role_id, status_code, client, auth_as_admin):
+    response = client.get(f"/auth/roles/{role_id}", headers=auth_as_admin)
     assert response.status_code == status_code
 
 
-def test_update_role(client, user_role_id):
-    response = client.put(f"/auth/roles/{user_role_id}", json={"name": "new name"})
+def test_update_role(client, user_role_id, auth_as_admin):
+    response = client.put(f"/auth/roles/{user_role_id}", json={"name": "new name"}, headers=auth_as_admin)
     assert response.status_code == HTTPStatus.NO_CONTENT
 
-    response = client.get(f"/auth/roles/{user_role_id}")
+    response = client.get(f"/auth/roles/{user_role_id}", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.OK
     assert response.json.get("name") == "new name"
 
@@ -86,17 +86,17 @@ def test_update_role(client, user_role_id):
         ("non_exist_uuid", {"name": "new_name"}, HTTPStatus.NOT_FOUND),
     ],
 )
-def test_update_role_errors(role_id_fixture, query, status_code, client, request):
+def test_update_role_errors(role_id_fixture, query, status_code, client, auth_as_admin, request):
     role_id = request.getfixturevalue(role_id_fixture)
-    response = client.put(f"/auth/roles/{role_id}", json=query)
+    response = client.put(f"/auth/roles/{role_id}", json=query, headers=auth_as_admin)
     assert response.status_code == status_code
 
 
-def test_delete_role(client, user_role_id):
-    response = client.delete(f"/auth/roles/{user_role_id}")
+def test_delete_role(client, user_role_id, auth_as_admin):
+    response = client.delete(f"/auth/roles/{user_role_id}", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.NO_CONTENT
 
-    response = client.get(f"/auth/roles/{user_role_id}")
+    response = client.get(f"/auth/roles/{user_role_id}", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.NOT_FOUND
 
 
@@ -107,19 +107,19 @@ def test_delete_role(client, user_role_id):
         ("7f32cd4a-7981-436d-bba7-78169acbbb5d", HTTPStatus.NOT_FOUND),
     ],
 )
-def test_delete_role_errors(role_id, status_code, client):
-    response = client.delete(f"/auth/roles/{role_id}")
+def test_delete_role_errors(role_id, status_code, client, auth_as_admin):
+    response = client.delete(f"/auth/roles/{role_id}", headers=auth_as_admin)
     assert response.status_code == status_code
 
 
-def test_no_default_role(client, example_user_id):
-    response = client.get(f"/auth/users/{example_user_id}/roles")
+def test_no_default_role(client, example_user_id, auth_as_admin):
+    response = client.get(f"/auth/users/{example_user_id}/roles", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.OK
     assert response.json == []
 
 
-def test_get_all_roles_of_user(client, example_with_roles_user_id):
-    response = client.get(f"/auth/users/{example_with_roles_user_id}/roles")
+def test_get_all_roles_of_user(client, example_with_roles_user_id, auth_as_admin):
+    response = client.get(f"/auth/users/{example_with_roles_user_id}/roles", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.OK
     assert len(response.json) == 3
     assert set(role["name"] for role in response.json) == {"user", "subscriber", "admin"}
@@ -132,18 +132,18 @@ def test_get_all_roles_of_user(client, example_with_roles_user_id):
         ("7f32cd4a-7981-436d-bba7-78169acbbb5d", HTTPStatus.NOT_FOUND),
     ],
 )
-def ttest_get_all_roles_of_user_errors(user_id, status_code, client):
-    response = client.get(f"/auth/users/{user_id}/roles")
+def test_get_all_roles_of_user_errors(user_id, status_code, client, auth_as_admin):
+    response = client.get(f"/auth/users/{user_id}/roles", headers=auth_as_admin)
     assert response.status_code == status_code
 
 
-def test_add_role_to_user(client, example_user_id, user_role_id):
-    response = client.post(f"/auth/users/{example_user_id}/roles/{user_role_id}")
+def test_add_role_to_user(client, example_user_id, user_role_id, auth_as_admin):
+    response = client.post(f"/auth/users/{example_user_id}/roles/{user_role_id}", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.CREATED
     assert len(response.json) == 1
     assert response.json[0]["name"] == "user"
 
-    response = client.get(f"/auth/users/{example_user_id}/roles")
+    response = client.get(f"/auth/users/{example_user_id}/roles", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.OK
     assert len(response.json) == 1
     assert response.json[0]["name"] == "user"
@@ -158,15 +158,15 @@ def test_add_role_to_user(client, example_user_id, user_role_id):
         ("non_exist_uuid", "user_role_id", HTTPStatus.NOT_FOUND),
     ],
 )
-def test_add_role_to_user_errors(user_id_fixture, role_id_fixture, status_code, client, request):
+def test_add_role_to_user_errors(user_id_fixture, role_id_fixture, status_code, client, auth_as_admin, request):
     user_id = request.getfixturevalue(user_id_fixture)
     role_id = request.getfixturevalue(role_id_fixture)
-    response = client.post(f"/auth/users/{user_id}/roles/{role_id}")
+    response = client.post(f"/auth/users/{user_id}/roles/{role_id}", headers=auth_as_admin)
     assert response.status_code == status_code
 
 
-def test_delete_role_from_user(client, example_with_roles_user_id, user_role_id):
-    response = client.delete(f"/auth/users/{example_with_roles_user_id}/roles/{user_role_id}")
+def test_delete_role_from_user(client, example_with_roles_user_id, auth_as_admin, user_role_id):
+    response = client.delete(f"/auth/users/{example_with_roles_user_id}/roles/{user_role_id}", headers=auth_as_admin)
     assert response.status_code == HTTPStatus.OK
     assert len(response.json) == 2
     assert set(role["name"] for role in response.json) == {"subscriber", "admin"}
@@ -181,8 +181,8 @@ def test_delete_role_from_user(client, example_with_roles_user_id, user_role_id)
         ("non_exist_uuid", "user_role_id", HTTPStatus.NOT_FOUND),
     ],
 )
-def test_delete_role_from_user_errors(user_id_fixture, role_id_fixture, status_code, client, request):
+def test_delete_role_from_user_errors(user_id_fixture, role_id_fixture, status_code, client, auth_as_admin, request):
     user_id = request.getfixturevalue(user_id_fixture)
     role_id = request.getfixturevalue(role_id_fixture)
-    response = client.delete(f"/auth/users/{user_id}/roles/{role_id}")
+    response = client.delete(f"/auth/users/{user_id}/roles/{role_id}", headers=auth_as_admin)
     assert response.status_code == status_code
